+++ package/src/firebase_auth/server/use_auth_user_server.ts
@@ -1,5 +1,4 @@
 import type { User } from 'firebase/auth';
-import { cache } from 'react';
 import { getAuthenticatedAppForUser } from './firebase_server';
 
 /**